refactor: Rename component files to PascalCase convention
Changes:
- Rename smallUserCard.tsx → SmallUserCard.tsx
- Rename smallAlbumCard.tsx → SmallAlbumCard.tsx
- Rename smallArtistCard.tsx → SmallArtistCard.tsx
- Update imports in searchResultClient.tsx

Result: All component files now follow consistent PascalCase naming.
Easier discovery and follows React/TypeScript conventions.
