+++ src/app/components/SmallAlbumCard.tsx
