+++ src/app/components/SmallArtistCard.tsx
